Test: add make script for rspec tests; divide physical device and simulator rspec tests
