+++ spec/physical_device_app_life_cycle_spec.rb
