+++ spec/simulator_set_location_spec.rb
