+++ spec/spec_helper.rb
@@ -4,6 +4,7 @@ require 'pry'
 require 'tmpdir'
 require 'fileutils'
 require "resources.rb"
+require "test_helper.rb"
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
